[add + change] Fragment上に移行できた
